requirements.txt
pillow==11.1.0
schedule==1.2.2
